split out newBasicAuthClient
