+++ pkg/sources/github/github.go
@@ -466,11 +466,7 @@ RepoLoop:
 }
 
 func (s *Source) enumerateBasicAuth(ctx context.Context, apiEndpoint string, basicAuth *credentialspb.BasicAuth) error {
-	s.httpClient.Transport = &github.BasicAuthTransport{
-		Username: basicAuth.Username,
-		Password: basicAuth.Password,
-	}
-	ghClient, err := createGitHubClient(s.httpClient, apiEndpoint)
+	ghClient, err := newBasicAuthClient(basicAuth, apiEndpoint)
 	if err != nil {
 		s.log.Error(err, "error creating GitHub client")
 	}
@@ -673,16 +669,6 @@ func (s *Source) enumerateWithApp(ctx context.Context, apiEndpoint string, app *
 	}
 
 	return installationClient, nil
-}
-
-func createGitHubClient(httpClient *http.Client, apiEndpoint string) (*github.Client, error) {
-	// If we're using public GitHub, make a regular client.
-	// Otherwise, make an enterprise client.
-	if strings.EqualFold(apiEndpoint, cloudEndpoint) {
-		return github.NewClient(httpClient), nil
-	}
-
-	return github.NewClient(httpClient).WithEnterpriseURLs(apiEndpoint, apiEndpoint)
 }
 
 func (s *Source) scan(ctx context.Context, installationClient *github.Client, chunksChan chan *sources.Chunk) error {
